confirurando rotas e auth

+++ front/src/app/guards/auth.guard.ts
@@ -8,23 +8,44 @@ export class AuthGuard implements CanActivate {
   constructor(private authService: AuthService, private router: Router) {}
 
   public canActivate(route: ActivatedRouteSnapshot): Observable<boolean> {
-    const permissoesAcesso = route.data['permissao'] as Array<number>;
+    const permissoesAcesso = route.data['permissoes'] as Array<number>;
+    const rotaLogin = route.routeConfig?.path === 'auth';
 
     return this.authService.usuarioLogado$.pipe(
       take(1),
       map((usuario) => {
-        if (!usuario) {
-          this.authService.removerAcesso();
-          this.router.navigate(['/auth']);
+        if (rotaLogin && usuario) {
+          this.redirecionarBaseadoNoNivelAcesso();
           return false;
         }
 
-        if (permissoesAcesso.includes(usuario.nivelAcesso)) {
-          return true;
+        if (!usuario && !permissoesAcesso) {
+          this.authService.removerAcesso();
+          this.router.navigate(['/dashboard']);
+          return false;
         }
 
-        return false;
+        if (usuario && permissoesAcesso) {
+          if (permissoesAcesso.includes(usuario.nivelAcesso)) {
+            return true;
+          }
+          this.redirecionarBaseadoNoNivelAcesso();
+          return false;
+        }
+        return true;
       })
     );
+  }
+
+  private redirecionarBaseadoNoNivelAcesso(): void {
+    if (this.authService.isAdmin()) {
+      this.router.navigate(['/dashboard/administrador-listagem']);
+    } else if (this.authService.isGeralLider()) {
+      this.router.navigate(['/dashboard/lideranca-listagem']);
+    } else if (this.authService.isGeralMembro()) {
+      this.router.navigate(['/dashboard/membro-listagem']);
+    } else {
+      this.authService.logout();
+    }
   }
 }
